Refactor popup into pages and global css
