+++ src/pages/Popup.svelte
@@ -1,0 +1,8 @@
+<script>
+  import { onMount } from "svelte";
+  import Summary from "./Summary.svelte";
+</script>
+
+<main>
+  <Summary />
+</main>
