+++ src/popup.js
@@ -1,4 +1,4 @@
-import App from "./components/Popup.svelte";
+import App from "./pages/Popup.svelte";
 
 const app = new App({
   target: document.body,
