+++ static/popup/global.css
@@ -1,0 +1,20 @@
+body {
+  background-image: url("https://maplelegends.com/static/images/bg.png");
+  background-position: center bottom;
+}
+
+main {
+  background-color: #ffffff;
+  opacity: 0.7;
+  padding: 0.25em 1em;
+  font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto,
+    Oxygen-Sans, Ubuntu, Cantarell, "Helvetica Neue", sans-serif;
+}
+
+h1 {
+  text-align: center;
+}
+
+h1 a {
+  text-decoration: none;
+}
