+++ static/popup/index.html
@@ -6,6 +6,7 @@
 
     <title>Popup</title>
 
+    <link rel="stylesheet" href="global.css" />
     <link rel="stylesheet" href="bundle.css" />
     <script defer src="bundle.js"></script>
   </head>
